extract visualization plumbing into utils.py
Move 11 helper functions out of solution.py into a new utils.py:
- TensorBoard plumbing: find_free_port, launch_tensorboard,
  log_batch_tensorboard, log_batch_jupyter
- Visualization post-processing: process_image
- RGB composite helpers: Color, BOP_ORANGE/BOP_BLUE/GREEN/MAGENTA,
  rescale_clip, composite_nuc_mem, clip_p, clip_highlight

Keep inline (pedagogically interesting):
- normalize_fov, min_max_scale (students should see how the input is
  normalized for the model vs how arrays are scaled for SSIM)
- feature_map_pca, pcs_to_rgb (used in the Part 2 visualization task)
- cellpose_segmentation (students should see what Cellpose is doing)

Notebook is now ~5% smaller (137 KB -> 130 KB for solution.ipynb).
The notebook entry point becomes a single `from utils import ...`
block with a comment telling curious students where the plumbing lives.

Side fixes: log_batch_jupyter now takes channel_names as an argument
instead of closing over a global. Drop unused `import torchvision` and
`from typing import NamedTuple`.
